20/10 Change Client Http from Rest Template

+++ src/main/java/bank/deposit/service/AccountDepositService.java
@@ -15,7 +15,7 @@ public interface AccountDepositService {
     Optional<AccountDepositDTO> getByIban(String iban);
     List<AccountDepositDTO> getAllByIndividual(int individualId);
     void deleteAccountDepositByIban(String iban);
-    AccountDepositDTO createIndividualAccountDeposit(int individualId, int months, int amount);
+    AccountDepositDTO createIndividualAccountDeposit(int individualId, int months, double amount);
     //AccountDepositDTO creditAccountDeposit(String iban, Double amount);
